+++ WiredBrainCoffee.ShopInfoTool/ICommandHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace WiredBrainCoffee.ShopInfoTool
+{
+    internal interface ICommandHandler
+    {
+        void HandleCommand();
+    }
+}
