+++ WiredBrainCoffee.ShopInfoTool/Program.cs
@@ -26,42 +26,12 @@ namespace WiredBrainCoffee.ShopInfoTool
 
                 var coffeeShops = coffeeShopDataProvider.LoadCoffeeShops();
 
-                if (string.Equals("help", line, StringComparison.OrdinalIgnoreCase))
-                {
-                    Console.WriteLine("> Available coffee shop commands:");
-                    foreach (var coffeeShop in coffeeShops)
-                    {
-                        Console.WriteLine($"> " + coffeeShop.Location);
-                    }
-                }
-                else
-                {
-                    var foundCoffeShops = coffeeShops
-                        .Where(x => x.Location.StartsWith(line, StringComparison.OrdinalIgnoreCase))
-                        .ToList();
+                var commandHandler =
+                    string.Equals("help", line, StringComparison.OrdinalIgnoreCase) 
+                    ? new HelpCommandHandler(coffeeShops) as ICommandHandler
+                    : new  CoffeeShopCommandHandler(coffeeShops, line);
 
-                    if (foundCoffeShops.Count == 0)
-                    {
-                        Console.WriteLine($"> Command '{line}' not found");
-                    }
-                    else if (foundCoffeShops.Count == 1)
-                    {
-                        var coffeeShop = foundCoffeShops.Single();
-                        Console.WriteLine($"> Location: {coffeeShop.Location}");
-                        Console.WriteLine($"> Beans in stock: {coffeeShop.BeansInStockInKg} kg");
-                    }
-                    else
-                    {
-                        Console.WriteLine($"> Multiple matching coffee shop commands found: ");
-                        foreach (var coffeeType in foundCoffeShops)
-                        {
-                            Console.WriteLine($"> {coffeeType.Location}");
-                        }
-                    }
-
-                    
-
-                }
+                commandHandler.HandleCommand();
             }
 
         }
